Shape cast fixes
Builds on Xcode now, but premake has a bug with cppdialect.

+++ Box2D/Box2D/Common/b2Timer.h
@@ -42,8 +42,8 @@ private:
 	float64 m_start;
 	static float64 s_invFrequency;
 #elif defined(__linux__) || defined (__APPLE__)
-	unsigned long m_start_sec;
-	unsigned long m_start_usec;
+	unsigned long long m_start_sec;
+	unsigned long long m_start_usec;
 #endif
 };
